basic routes and their documentation
